Feat: maxValue 타입변경

+++ src/main/java/com/mycompany/metamong/dto/sequence/SequenceDto.java
@@ -1,5 +1,7 @@
 package com.mycompany.metamong.dto.sequence;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -7,7 +9,7 @@ public class SequenceDto {
 	private String sequenceName;
 	private String schemaName;
 	private int minValue;
-	private int maxValue;
+	private BigDecimal maxValue;
 	private int incrementBy;
 	private int lastNumber;
